int64

+++ MultiplicationOfNumbers/Program.cs
@@ -1,7 +1,7 @@
 ﻿Console.WriteLine("Enter a limit for calculating multiplication from 1:");
-int limit=int.Parse (Console.ReadLine());
+Int64 limit=Int64.Parse (Console.ReadLine());
 int count = 1;
-int multi=1;
+Int64 multi=1;
 while (count<=limit)
     {
     multi=multi*count;
